Fixed bug with reseting color dialog in varialbe color property. Added abbility to change varialbe class.

+++ src/smartprop_editor/properties_classes/float.py
@@ -64,7 +64,7 @@ class PropertyFloat(QWidget):
 
 
         # Variable setup
-        self.variable = ComboboxVariablesWidget(variables_layout=self.variables_scrollArea, filter_types=['Int', 'Float'])
+        self.variable = ComboboxVariablesWidget(variables_layout=self.variables_scrollArea, filter_types=['Float', 'Int'])
         layout = QHBoxLayout()
         layout.setContentsMargins(0, 0, 0, 0)
         layout.addWidget(self.variable)
